requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
python-dotenv==1.0.1
psycopg2-binary==2.9.10
# Windows-only helper (ignored on Linux/Vercel):
pywin32>=305; platform_system == "Windows"
